[Upd #286] Refactor to remove circular dependency of imports
